♻️ Refactor mock utils
