+++ modules/fantasy-battle/src/Commands/bio.test.ts
@@ -1,7 +1,7 @@
 import { mockMessage } from "@discord-bot/discord-mock"
 import PcModel from "Models/PlayerCharacter"
 import PlayerUserModel from "Models/PlayerUser"
-import { mockDmMessage, mockPlayerMessage } from "Utils/mockMessage"
+import { mockDmMessage, mockPlayerMessage } from "Utils/Mock/mockMessage"
 import { useDbConnection } from "Utils/Mongo/mongoTest"
 
 import { test, execute } from "./bio"
