+++ modules/fantasy-battle/src/Commands/listChars.test.ts
@@ -3,7 +3,7 @@ import { mockMessage } from "@discord-bot/discord-mock"
 import PcModel from "Models/PlayerCharacter"
 import PlayerUserModel from "Models/PlayerUser"
 
-import { mockPlayerMessage } from "Utils/mockMessage"
+import { mockPlayerMessage } from "Utils/Mock/mockMessage"
 import { useDbConnection } from "Utils/Mongo/mongoTest"
 
 import { test, execute } from "./listChars"
