+++ modules/fantasy-battle/src/Models/PlayerCharacter/index.test.ts
@@ -1,5 +1,5 @@
 import ResourceModel from "Models/PcResource"
-import mockAttributes from "Utils/mockAttributes"
+import mockAttributes from "Utils/Mock/mockAttributes"
 import { useDbConnection } from "Utils/Mongo/mongoTest"
 import PcModel, { Pc } from "./index"
 import { createPcProps } from "./statics/create"
