+++ modules/fantasy-battle/src/Models/PlayerCharacter/methods/rollDmg.test.ts
@@ -1,4 +1,4 @@
-import mockAttributes from "Utils/mockAttributes"
+import mockAttributes from "Utils/Mock/mockAttributes"
 import PcModel, { Pc } from ".."
 import rollDmg from "./rollDmg"
