+++ modules/fantasy-battle/src/Utils/Mock/mockMessage.ts
